test: isolate migration validation from backend suite
