+++ pytest.ini
@@ -12,8 +12,11 @@ minversion = 7.0
 # Async support
 asyncio_mode = auto
 
-# Test paths
-testpaths = test
+# The backend suite exercises application code under ``src/kohakuhub``.
+# Migration-history validation lives in ``test/migrations`` and is invoked
+# explicitly by ``make test-migrations`` so it cannot be mistaken for a unit
+# test or become part of an ordinary ``pytest`` run.
+testpaths = test/kohakuhub
 
 # Output options
 addopts =
